requirements.txt
astroid==2.4.2
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.1
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Scrypt==0.1.3.6
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fpdf==1.7.2
gunicorn==20.0.4
idna==2.8
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
limits==1.5.1
lml==0.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.19
pyecharts-jupyter-installer==0.0.3
pylint==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.22.0
scrypt==0.8.17
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.13
texttable==1.6.2
toml==0.10.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.3.3
xlrd==1.2.0
xlwt==1.3.0
